mock static method of final class NetworkInterface by Mockito
